For MAF intergration
The largest difference is that the query to build the list of blending stars is external and is made in the notebook (see for example Notebook_MAF.ipynb). If no internet connection or if you don't have  TRILEGAL lsst_sim.simdr2 df is empty and the metric works. The output is a complex dictionary.